- use INTEGER on oracle instead of INTEGER(x): it is faster and uses the same space anyway, plus it is coherent with rest of db
git-svn-id: http://svn.projects.ez.no/ezsi/trunk@52 0cc08a9f-4d52-4612-8d77-e0d2f54f1ed6

+++ sql/oracle/schema.sql
@@ -1,8 +1,8 @@
 CREATE TABLE ezsi_files (
     filepath VARCHAR2(4000) NOT NULL,
     namehash CHAR(32) PRIMARY KEY,
-    mtime INTEGER(11) NOT NULL DEFAULT 0,
+    mtime INTEGER NOT NULL DEFAULT 0,
     urlalias VARCHAR2(4000) NOT NULL,
     siteaccess VARCHAR2(100) NOT NULL,
-    ttl INTEGER(10) NOT NULL,
+    ttl INTEGER NOT NULL,
     blockkeys VARCHAR2(4000) NOT NULL);
